feat: implement application workflow, user management, and location hierarchy systems with supporting backend modules and frontend interfaces
