+++ frontend/src/app/admin/users/[id]/edit/page.tsx
@@ -1,205 +1,10 @@
-"use client";
+import React from 'react';
+import EditUserPage from './EditUserClient';
 
-import { useEffect, useState, Suspense } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
-import { AdminFormSkeleton } from '@/components/admin';
+export async function generateStaticParams() {
+  return [{ id: '1' }];
+}
 
-const roles = ['Admin', 'User', 'Manager'];
-
-// Component that uses useSearchParams - needs to be wrapped in Suspense
-// Skeleton component for the edit form page
-const EditFormSkeleton = () => (
-  <div className="flex flex-col items-center justify-center min-h-screen bg-gray-50">
-    <div className="bg-white p-8 rounded shadow-md w-full max-w-md">
-      <div className="h-8 w-40 bg-gray-200 rounded mb-6 animate-pulse"></div>
-      <AdminFormSkeleton fields={5} showButton={true} />
-    </div>
-  </div>
-);
-
-const EditUserContent = () => {
-  const [username, setUsername] = useState('');
-  const [email, setEmail] = useState('');
-  const [phone, setPhone] = useState('');
-  const [password, setPassword] = useState('');
-  const [role, setRole] = useState(roles[0]);
-  const [policeStationId, setPoliceStationId] = useState('');
-  const [stateId, setStateId] = useState('');
-  const [districtId, setDistrictId] = useState('');
-  const [zoneId, setZoneId] = useState('');
-  const [divisionId, setDivisionId] = useState('');
-  const [error, setError] = useState<string | null>(null);
-  const [success, setSuccess] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const userId = searchParams?.get('id');
-
-  useEffect(() => {
-    const fetchUserDetails = async () => {
-      try {
-        setIsLoading(true);
-        setError(null);
-
-        const response = await fetch(`/admin/users/${userId}`);
-        if (!response.ok) {
-          throw new Error('Failed to fetch user details');
-        }
-
-        const data = await response.json();
-        setUsername(data.username);
-        setEmail(data.email);
-        setPhone(data.phone);
-        setRole(data.role);
-      } catch (err) {
-        setError(err instanceof Error ? err.message : 'An unknown error occurred');
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    if (userId) {
-      fetchUserDetails();
-    }
-  }, [userId]);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError(null);
-    setSuccess(null);
-    setIsSubmitting(true);
-
-    try {
-      const response = await fetch(`/admin/users/${userId}`, {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ username, email, phone, password:password || undefined, roleCode: role, policeStationId, stateId:Number(stateId), districtId:Number(districtId), zoneId:Number(zoneId), divisionId:Number(divisionId)}),
-      });
-
-      if (!response.ok) {
-        throw new Error('Failed to update user');
-      }
-
-      setSuccess('User updated successfully!');
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'An unknown error occurred');
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  if (!userId) {
-    return <div>User ID is missing</div>;
-  }
-
-  if (isLoading) {
-    return (
-      <div className="flex flex-col items-center justify-center min-h-screen bg-gray-50">
-        <div className="bg-white p-8 rounded shadow-md w-full max-w-md">
-          <div className="h-8 w-40 bg-gray-200 rounded mb-6 animate-pulse"></div>
-          <AdminFormSkeleton fields={5} showButton={true} />
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex flex-col items-center justify-center min-h-screen bg-gray-50">
-      <form onSubmit={handleSubmit} className="bg-white p-8 rounded shadow-md w-full max-w-md">
-        <h1 className="text-2xl font-bold mb-4">Edit User</h1>
-
-        {error && <div className="mb-4 text-red-600">{error}</div>}
-        {success && <div className="mb-4 text-green-600">{success}</div>}
-
-        <div className="mb-4">
-          <label className="block mb-1">Username</label>
-          <input
-            type="text"
-            value={username}
-            onChange={(e) => setUsername(e.target.value)}
-            required
-            disabled={isSubmitting}
-            className="w-full border rounded px-3 py-2 disabled:bg-gray-100"
-          />
-        </div>
-
-        <div className="mb-4">
-          <label className="block mb-1">Email</label>
-          <input
-            type="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-            disabled={isSubmitting}
-            className="w-full border rounded px-3 py-2 disabled:bg-gray-100"
-          />
-        </div>
-
-        <div className="mb-4">
-          <label className="block mb-1">Phone</label>
-          <input
-            type="tel"
-            value={phone}
-            onChange={(e) => setPhone(e.target.value)}
-            required
-            disabled={isSubmitting}
-            className="w-full border rounded px-3 py-2 disabled:bg-gray-100"
-          />
-        </div>
-
-        <div className="mb-4">
-          <label className="block mb-1">Password</label>
-          <input
-            type="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            disabled={isSubmitting}
-            className="w-full border rounded px-3 py-2 disabled:bg-gray-100"
-          />
-        </div>
-
-        <div className="mb-4">
-          <label className="block mb-1">Role</label>
-          <select
-            value={role}
-            onChange={(e) => setRole(e.target.value)}
-            disabled={isSubmitting}
-            className="w-full border rounded px-3 py-2 disabled:bg-gray-100"
-          >
-            {roles.map((r) => (
-              <option key={r} value={r}>
-                {r}
-              </option>
-            ))}
-          </select>
-        </div>
-
-        <button
-          type="submit"
-          disabled={isSubmitting}
-          className="w-full bg-blue-600 text-white py-2 rounded hover:bg-blue-700 disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center gap-2"
-        >
-          {isSubmitting && (
-            <svg className='animate-spin h-5 w-5 text-white' xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 24 24'>
-              <circle className='opacity-25' cx='12' cy='12' r='10' stroke='currentColor' strokeWidth='4' />
-              <path className='opacity-75' fill='currentColor' d='M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z' />
-            </svg>
-          )}
-          {isSubmitting ? 'Updating...' : 'Update User'}
-        </button>
-      </form>
-    </div>
-  );
-};
-
-// Main component with Suspense boundary
-const EditUserPage = () => {
-  return (
-    <Suspense fallback={<EditFormSkeleton />}>
-      <EditUserContent />
-    </Suspense>
-  );
-};
-
-export default EditUserPage;
+export default async function Page() {
+  return <EditUserPage />;
+}
