+++ frontend/src/types/location.ts
@@ -12,12 +12,18 @@ export interface State extends LocationOption {
 export interface District extends LocationOption {
   stateId: number;
   state?: State;
+  rangeOffices?: RangeOffice[];
+}
+
+export interface RangeOffice extends LocationOption {
+  districtId: number;
+  district?: District;
   zones?: Zone[];
 }
 
 export interface Zone extends LocationOption {
-  districtId: number;
-  district?: District;
+  rangeOfficeId?: number;
+  rangeOffice?: RangeOffice;
   divisions?: Division[];
 }
 
@@ -44,11 +50,13 @@ export interface LocationAPIResponse<T> {
 export interface AddressLocationData {
   state: string;
   district: string;
+  rangeOffice: string;
   zone: string;
   division: string;
   policeStation: string;
   stateName?: string;
   districtName?: string;
+  rangeOfficeName?: string;
   zoneName?: string;
   divisionName?: string;
   policeStationName?: string;
@@ -58,11 +66,13 @@ export interface AddressFormData {
   presentAddress: string;
   presentState: string;
   presentDistrict: string;
+  presentRangeOffice: string;
   presentZone: string;
   presentDivision: string;
   presentPoliceStation: string;
   presentStateName?: string;
   presentDistrictName?: string;
+  presentRangeOfficeName?: string;
   presentZoneName?: string;
   presentDivisionName?: string;
   presentPoliceStationName?: string;
@@ -71,11 +81,13 @@ export interface AddressFormData {
   permanentAddress: string;
   permanentState: string;
   permanentDistrict: string;
+  permanentRangeOffice: string;
   permanentZone: string;
   permanentDivision: string;
   permanentPoliceStation: string;
   permanentStateName?: string;
   permanentDistrictName?: string;
+  permanentRangeOfficeName?: string;
   permanentZoneName?: string;
   permanentDivisionName?: string;
   permanentPoliceStationName?: string;
@@ -89,16 +101,19 @@ export interface AddressFormData {
 export interface LocationHierarchyState {
   states: State[];
   districts: District[];
+  rangeOffices: RangeOffice[];
   zones: Zone[];
   divisions: Division[];
   policeStations: PoliceStation[];
   selectedState: string;
   selectedDistrict: string;
+  selectedRangeOffice: string;
   selectedZone: string;
   selectedDivision: string;
   selectedPoliceStation: string;
   loadingStates: boolean;
   loadingDistricts: boolean;
+  loadingRangeOffices: boolean;
   loadingZones: boolean;
   loadingDivisions: boolean;
   loadingPoliceStations: boolean;
